+++ ROItem/service/accnameService.py
@@ -22,4 +22,9 @@ def accname_create(Accname):
 
 def delete_Accname_Table(view_name):
     createData = models.Accname.objects
-    createData.filter(view_name__in=view_name).delete()
+    if not view_name:
+        return
+    if isinstance(view_name, (list, tuple, set)):
+        createData.filter(view_name__in=view_name).delete()
+    else:
+        createData.filter(view_name=view_name).delete()
